Convert error message from bytes to text
The $error veriable has to be utf8 decoded.
Otherwise non-ascii characters would appear broken.

Since this mdule is doing a `use bytes` at the beginning, we have
to do that expicitly.

That's also why the input `$keys` is received utf8 encoded, although
the calling function passes it decoded.

Using `bytes` is actually discouraged, and maybe we can get rid of it and
replace it locally with what is actually needed:
https://perldoc.perl.org/bytes

Issue: https://progress.opensuse.org/issues/133451